+++ src/miiservice/listfoldersservice.ts
@@ -4,7 +4,7 @@ import { Folder, GeneralColumn2, MII } from './abstract/responsetypes';
 
 class ListFoldersService extends Service {
     name: string = "List Folders";
-    mode: string = "XMII/Catalog?Mode=ListFolders&Session=true&DoStateCheck=true&Content-Type=text/xml";
+    mode: string = "XMII/Catalog?Mode=ListFolders&DoStateCheck=true&Content-Type=text/xml";
     async call({ host, port }: Request, folderPath: string) {
         const url = this.get(host, port, folderPath);
         const { value, error, isError } = await this.fetch(new URL(url));